Abandon dynamic route loading
