+++ dallinger/experiment_server/experiment_server.py
@@ -1,26 +1,14 @@
 """ This module provides the backend Flask server that serves an experiment. """
 
-from datetime import datetime, timedelta
-from functools import update_wrapper
+from datetime import datetime
 import gevent
 from json import dumps
 from json import loads
-from operator import attrgetter
 import os
 import re
-import sys
 
-from flask import (
-    abort,
-    Flask,
-    make_response,
-    render_template,
-    request,
-    Response,
-    send_from_directory,
-)
+from flask import abort, Flask, render_template, request, Response, send_from_directory
 from jinja2 import TemplateNotFound
-from rq import get_current_job
 from rq import Queue
 from sqlalchemy.orm.exc import NoResultFound, MultipleResultsFound
 from sqlalchemy import exc
@@ -28,20 +16,25 @@ from sqlalchemy import func
 from sqlalchemy.sql.expression import true
 from psycopg2.extensions import TransactionRollbackError
 
-
 from dallinger import db
 from dallinger import experiment
 from dallinger import models
-from dallinger import information
 from dallinger.config import get_config
 from dallinger import recruiters
 from dallinger.notifications import get_messenger
 from dallinger.notifications import MessengerError
 
 from .replay import ReplayBackend
-from .worker_events import WorkerEvent
-from .utils import nocache
-from .utils import ValidatesBrowser
+from .worker_events import worker_function
+from .utils import (
+    crossdomain,
+    nocache,
+    ValidatesBrowser,
+    error_page,
+    error_response,
+    success_response,
+    ExperimentError,
+)
 
 
 # Initialize the Dallinger database.
@@ -53,54 +46,6 @@ q = Queue(connection=redis_conn)
 WAITING_ROOM_CHANNEL = "quorum"
 
 app = Flask("Experiment_Server")
-
-
-def crossdomain(
-    origin=None,
-    methods=None,
-    headers=None,
-    max_age=21600,
-    attach_to_all=True,
-    automatic_options=True,
-):
-    if methods is not None:
-        methods = ", ".join(sorted(x.upper() for x in methods))
-    if headers is not None and not isinstance(headers, str):
-        headers = ", ".join(x.upper() for x in headers)
-    if not isinstance(origin, str):
-        origin = ", ".join(origin)
-    if isinstance(max_age, timedelta):
-        max_age = max_age.total_seconds()
-
-    def get_methods():
-        if methods is not None:
-            return methods
-
-        options_resp = app.make_default_options_response()
-        return options_resp.headers["allow"]
-
-    def decorator(f):
-        def wrapped_function(*args, **kwargs):
-            if automatic_options and request.method == "OPTIONS":
-                resp = app.make_default_options_response()
-            else:
-                resp = make_response(f(*args, **kwargs))
-            if not attach_to_all and request.method != "OPTIONS":
-                return resp
-
-            h = resp.headers
-
-            h["Access-Control-Allow-Origin"] = origin
-            h["Access-Control-Allow-Methods"] = get_methods()
-            h["Access-Control-Max-Age"] = str(max_age)
-            if headers is not None:
-                h["Access-Control-Allow-Headers"] = headers
-            return resp
-
-        f.provide_automatic_options = False
-        return update_wrapper(wrapped_function, f)
-
-    return decorator
 
 
 @app.before_first_request
@@ -117,8 +62,7 @@ def Experiment(args):
     return klass(args)
 
 
-"""Load the experiment's extra routes, if any."""
-
+# Load the experiment's extra routes, if any.
 try:
     from dallinger_experiment.experiment import extra_routes
 except ImportError:
@@ -126,6 +70,10 @@ except ImportError:
 else:
     app.register_blueprint(extra_routes)
 
+# Ideally, we'd only load recruiter routes if the recruiter is active, but
+# it turns out this is complicated, so for now we always register our
+# primary recruiter's route:
+app.register_blueprint(recruiters.mturk_routes)
 
 """Basic routes."""
 
@@ -152,138 +100,6 @@ def static_robots_txt():
 @app.route("/favicon.ico")
 def static_favicon():
     return send_from_directory("static", "favicon.ico", mimetype="image/x-icon")
-
-
-"""Define some canned response types."""
-
-
-def success_response(**data):
-    """Return a generic success response."""
-    data_out = {}
-    data_out["status"] = "success"
-    data_out.update(data)
-    js = dumps(data_out, default=date_handler)
-    return Response(js, status=200, mimetype="application/json")
-
-
-def error_response(
-    error_type="Internal server error",
-    error_text="",
-    status=400,
-    participant=None,
-    simple=False,
-    request_data="",
-):
-    """Return a generic server error response."""
-    last_exception = sys.exc_info()
-    if last_exception[0]:
-        db.logger.error(
-            "Failure for request: {!r}".format(dict(request.args)),
-            exc_info=last_exception,
-        )
-
-    data = {"status": "error"}
-
-    if simple:
-        data["message"] = error_text
-    else:
-        data["html"] = (
-            error_page(
-                error_text=error_text,
-                error_type=error_type,
-                participant=participant,
-                request_data=request_data,
-            )
-            .get_data()
-            .decode("utf-8")
-        )
-    return Response(dumps(data), status=status, mimetype="application/json")
-
-
-def error_page(
-    participant=None,
-    error_text=None,
-    compensate=True,
-    error_type="default",
-    request_data="",
-):
-    """Render HTML for error page."""
-    config = _config()
-
-    if error_text is None:
-        error_text = """There has been an error and so you are unable to
-        continue, sorry!"""
-
-    if participant is not None:
-        hit_id = (participant.hit_id,)
-        assignment_id = (participant.assignment_id,)
-        worker_id = participant.worker_id
-        participant_id = participant.id
-    else:
-        hit_id = request.form.get("hit_id", "")
-        assignment_id = request.form.get("assignment_id", "")
-        worker_id = request.form.get("worker_id", "")
-        participant_id = request.form.get("participant_id", None)
-
-    if participant_id:
-        try:
-            participant_id = int(participant_id)
-        except (ValueError, TypeError):
-            participant_id = None
-
-    return make_response(
-        render_template(
-            "error.html",
-            error_text=error_text,
-            compensate=compensate,
-            contact_address=config.get("contact_email_on_error"),
-            error_type=error_type,
-            hit_id=hit_id,
-            assignment_id=assignment_id,
-            worker_id=worker_id,
-            request_data=request_data,
-            participant_id=participant_id,
-        ),
-        500,
-    )
-
-
-class ExperimentError(Exception):
-    """
-    Error class for experimental errors, such as subject not being found in
-    the database.
-    """
-
-    def __init__(self, value):
-        experiment_errors = dict(
-            status_incorrectly_set=1000,
-            hit_assign_worker_id_not_set_in_mturk=1001,
-            hit_assign_worker_id_not_set_in_consent=1002,
-            hit_assign_worker_id_not_set_in_exp=1003,
-            hit_assign_appears_in_database_more_than_once=1004,
-            already_started_exp=1008,
-            already_started_exp_mturk=1009,
-            already_did_exp_hit=1010,
-            tried_to_quit=1011,
-            intermediate_save=1012,
-            improper_inputs=1013,
-            browser_type_not_allowed=1014,
-            api_server_not_reachable=1015,
-            ad_not_found=1016,
-            error_setting_worker_complete=1017,
-            hit_not_registered_with_ad_server=1018,
-            template_unsafe=1019,
-            insert_mode_failed=1020,
-            page_not_found=404,
-            in_debug=2005,
-            unknown_error=9999,
-        )
-        self.value = value
-        self.errornum = experiment_errors[self.value]
-        self.template = "error.html"
-
-    def __str__(self):
-        return repr(self.value)
 
 
 @app.errorhandler(ExperimentError)
@@ -469,7 +285,6 @@ def launch():
         )
 
     try:
-        exp.recruiter.bind_app(app)
         recruitment_details = exp.recruiter.open_recruitment(
             n=exp.initial_recruitment_size
         )
@@ -1742,124 +1557,6 @@ def _worker_failed(participant_id):
             participant_id=participant.id,
             event_type="BotAssignmentRejected",
         )
-
-
-@db.scoped_session_decorator
-def worker_function(
-    event_type, assignment_id, participant_id, node_id=None, details=None
-):
-    """Process the notification."""
-    _config()
-    try:
-        db.logger.debug(
-            "rq: worker_function working on job id: %s", get_current_job().id
-        )
-        db.logger.debug(
-            "rq: Received Queue Length: %d (%s)", len(q), ", ".join(q.job_ids)
-        )
-    except AttributeError:
-        db.logger.debug("Debug worker_function called synchronously")
-
-    exp = Experiment(session)
-    key = "-----"
-
-    exp.log(
-        "Received an {} notification for assignment {}, participant {}".format(
-            event_type, assignment_id, participant_id
-        ),
-        key,
-    )
-
-    if event_type == "TrackingEvent":
-        node = None
-        if node_id:
-            node = models.Node.query.get(node_id)
-        if not node:
-            participant = None
-            if participant_id:
-                # Lookup assignment_id to create notifications
-                participant = models.Participant.query.get(participant_id)
-            elif assignment_id:
-                participants = models.Participant.query.filter_by(
-                    assignment_id=assignment_id
-                ).all()
-                # if there are one or more participants select the most recent
-                if participants:
-                    participant = max(participants, key=attrgetter("creation_time"))
-                    participant_id = participant.id
-            if not participant:
-                exp.log(
-                    "Warning: No participant associated with this "
-                    "TrackingEvent notification.",
-                    key,
-                )
-                return
-            nodes = participant.nodes()
-            if not nodes:
-                exp.log(
-                    "Warning: No node associated with this "
-                    "TrackingEvent notification.",
-                    key,
-                )
-                return
-            node = max(nodes, key=attrgetter("creation_time"))
-
-        if not details:
-            details = {}
-        info = information.TrackingEvent(origin=node, details=details)
-        session.add(info)
-        session.commit()
-        return
-
-    runner_cls = WorkerEvent.for_name(event_type)
-    if not runner_cls:
-        exp.log("Event type {} is not supported... ignoring.".format(event_type))
-        return
-
-    if assignment_id is not None:
-        # save the notification to the notification table
-        notif = models.Notification(assignment_id=assignment_id, event_type=event_type)
-        session.add(notif)
-        session.commit()
-
-        # try to identify the participant
-        participants = models.Participant.query.filter_by(
-            assignment_id=assignment_id
-        ).all()
-
-        # if there are one or more participants select the most recent
-        if participants:
-            participant = max(participants, key=attrgetter("creation_time"))
-
-        # if there are none print an error
-        else:
-            exp.log(
-                "Warning: No participants associated with this "
-                "assignment_id. Notification will not be processed.",
-                key,
-            )
-            return None
-
-    elif participant_id is not None:
-        participant = models.Participant.query.filter_by(id=participant_id).all()[0]
-    else:
-        raise ValueError(
-            "Error: worker_function needs either an assignment_id or a "
-            "participant_id, they cannot both be None"
-        )
-
-    participant_id = participant.id
-
-    runner = runner_cls(
-        participant, assignment_id, exp, session, _config(), datetime.now()
-    )
-    runner()
-    session.commit()
-
-
-def date_handler(obj):
-    """Serialize dates."""
-    return obj.isoformat() if hasattr(obj, "isoformat") else obj
 
 
 # Insert "mode" into pages so it's carried from page to page done server-side
